Version 2.6.2

+++ endpoint/store/include/privmx/endpoint/store/encryptors/fileData/ChunkEncryptor.hpp
@@ -26,13 +26,13 @@ class ChunkEncryptor : public IChunkEncryptor
 public:
     ChunkEncryptor(std::string key, size_t chunkSize);
     IChunkEncryptor::Chunk encrypt(const uint64_t index, const std::string& data) override;
-    std::string decrypt(const size_t index, const Chunk& chunk) override;
+    std::string decrypt(const uint64_t index, const Chunk& chunk) override;
     size_t getPlainChunkSize() override;
     size_t getEncryptedChunkSize() override;
     uint64_t getEncryptedFileSize(const uint64_t& fileSize) override;
     void sync(std::string key, size_t chunkSize) override;
 private:
-    std::string chunkIndexToBE(const size_t index);
+    std::string chunkIndexToBE(const uint64_t index);
 
     std::string _key;
     size_t _chunkSize;
